♻️ Extract assemble from startLogs
