+++ packages/logs/src/boot/startLogs.ts
@@ -32,6 +32,7 @@ import type { LogsSessionManager } from '../domain/logsSessionManager'
 import { startLogsSessionManager, startLogsSessionManagerStub } from '../domain/logsSessionManager'
 import type { LogsConfiguration } from '../domain/configuration'
 import type { LogsEvent } from '../logsEvent.types'
+import { buildAssemble, getRUMInternalContext } from '../domain/assemble'
 
 const LogStatusForApi = {
   [ConsoleApiName.log]: StatusType.info,
@@ -195,60 +196,4 @@ export function doStartLogs(
       onLogEventCollected(contextualizedMessage)
     }
   }
-}
-
-export function buildAssemble(
-  sessionManager: LogsSessionManager,
-  configuration: LogsConfiguration,
-  reportRawError: (error: RawError) => void
-) {
-  const logRateLimiters = {
-    [StatusType.error]: createEventRateLimiter(
-      StatusType.error,
-      configuration.eventRateLimiterThreshold,
-      reportRawError
-    ),
-    [StatusType.warn]: createEventRateLimiter(StatusType.warn, configuration.eventRateLimiterThreshold, reportRawError),
-    [StatusType.info]: createEventRateLimiter(StatusType.info, configuration.eventRateLimiterThreshold, reportRawError),
-    [StatusType.debug]: createEventRateLimiter(
-      StatusType.debug,
-      configuration.eventRateLimiterThreshold,
-      reportRawError
-    ),
-    ['custom']: createEventRateLimiter('custom', configuration.eventRateLimiterThreshold, reportRawError),
-  }
-
-  return (message: LogsMessage, currentContext: Context) => {
-    const startTime = message.date ? getRelativeTime(message.date) : undefined
-    const session = sessionManager.findTrackedSession(startTime)
-
-    if (!session) {
-      return undefined
-    }
-
-    const contextualizedMessage = combine(
-      { service: configuration.service, session_id: session.id },
-      currentContext,
-      getRUMInternalContext(startTime),
-      message
-    )
-
-    if (
-      configuration.beforeSend?.(contextualizedMessage) === false ||
-      (logRateLimiters[contextualizedMessage.status] ?? logRateLimiters['custom']).isLimitReached()
-    ) {
-      return undefined
-    }
-
-    return contextualizedMessage as Context
-  }
-}
-
-interface Rum {
-  getInternalContext: (startTime?: RelativeTime) => Context
-}
-
-function getRUMInternalContext(startTime?: RelativeTime): Context | undefined {
-  const rum = (window as any).DD_RUM as Rum
-  return rum && rum.getInternalContext ? rum.getInternalContext(startTime) : undefined
 }
